Refactor API function to use instanceID instead of Port
The `loadAPI` function in `apiFunctions.js` has been refactored to use the `instanceID` parameter instead of the `Port` parameter. This change allows for a more flexible and accurate way of loading the API. The commit also includes a minor update to the API endpoint URL.

Fix chat message sending in sendChat.js

In the `sendChat.js` file, the `sendConsoleMessage` function has been updated to send chat messages to the server with the correct formatting. The previous implementation used the `say` command, but it has been changed to use the `tellraw` command with the appropriate message format. This ensures that the chat messages are displayed correctly in the game.